Enhance XML protocol and add RXML utilities
- Updated `.gemini/settings.json` to enable checkpointing for improved testing.
- Refactored `morph-xml-protocol.ts` to streamline XML parsing and enhance error handling, including the introduction of new RXML utility functions for better modularity.
- Added `rxml.ts` to encapsulate XML parsing logic, including functions for extracting raw inner content, counting tag occurrences, and handling duplicate string tags.
- Improved argument processing in `morphXmlProtocol` to utilize the new RXML utilities, ensuring more robust handling of XML tool calls.

These changes contribute to a more maintainable and efficient XML protocol implementation in the parser package.
