+++ .gemini/settings.json
@@ -1,1 +1,8 @@
-{ "contextFileName": "AGENTS.md" }
+{
+  "contextFileName": "AGENTS.md",
+  "general": {
+    "checkpointing": {
+      "enabled": true
+    }
+  }
+}
