+++ packages/parser/src/utils/index.ts
@@ -2,6 +2,7 @@ import { createDynamicIfThenElseSchema } from "./dynamic-tool-schema";
 import { getPotentialStartIndex } from "./get-potential-start-index";
 import { escapeRegExp } from "./regex";
 import * as RJSON from "./relaxed-json";
+import * as RXML from "./rxml";
 export * from "./coercion";
 export * from "./debug";
 export * from "./on-error";
@@ -15,4 +16,5 @@ export {
   escapeRegExp,
   getPotentialStartIndex,
   RJSON,
+  RXML,
 };
